refactor(pool): add psycopg_pool.abc module
